+++ app/src/main/java/com/example/fragmentstest/models/MyViewHolder.kt
@@ -8,7 +8,7 @@ import kotlinx.android.synthetic.main.row_main.view.*
 class MyViewHolder(private val view: View) : RecyclerView.ViewHolder(view) {
 
     fun onBindViewHolder(user: User) {
-        view.circleImageView.setImageResource(user.photo)
+        view.circleImageView.setImageResource(user.photo.toInt())
         view.mr_tv_name.text = user.name
         view.mr_tv_number.text = user.number
         if (user.isFavorite)
